# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VAS v7.0",
    "version": "7.0.0"
  },
  "paths": {
...
    "/api/v1/upload": {
      "post": {
        "tags": [
          "assessment"
        ],
        "summary": "Upload Audio",
        "description": "上传并分析音频文件",
        "operationId": "upload_audio_api_v1_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_audio_api_v1_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyze": {
      "post": {
        "tags": [
          "assessment"
        ],
        "summary": "Analyze File",
        "description": "分析已存在的音频文件",
        "operationId": "analyze_file_api_v1_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/extract-pitch": {
      "post": {
        "tags": [
          "assessment"
        ],
        "summary": "Extract Pitch",
        "description": "提取音高曲线",
        "operationId": "extract_pitch_api_v1_extract_pitch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_extract_pitch_api_v1_extract_pitch_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/separate": {
      "post": {
        "tags": [
          "assessment"
        ],
        "summary": "Separate Audio",
        "description": "人声分离 (Demucs)",
        "operationId": "separate_audio_api_v1_separate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeparateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeparateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/report": {
      "post": {
        "tags": [
          "assessment"
        ],
        "summary": "Generate Report",
        "description": "生成评估报告 (PDF/图片)",
        "operationId": "generate_report_api_v1_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "properties": {
          "filepath": {
            "type": "string",
            "title": "Filepath"
          },
          "reference_filepath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Filepath"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "quick"
          }
        },
        "type": "object",
        "required": [
          "filepath"
        ],
        "title": "AnalyzeRequest",
        "description": "分析请求"
      },
      "Body_extract_pitch_api_v1_extract_pitch_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          }
        },
        "type": "object",
        "title": "Body_extract_pitch_api_v1_extract_pitch_post"
      },
      "Body_upload_audio_api_v1_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "quick"
          },
          "reference_file": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_audio_api_v1_upload_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportRequest": {
        "properties": {
          "analysis_result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Analysis Result"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "default": "report"
          },
          "format": {
            "type": "string",
            "title": "Format",
            "default": "image"
          }
        },
        "type": "object",
        "required": [
          "analysis_result"
        ],
        "title": "ReportRequest",
        "description": "报告生成请求"
      },
      "ReportResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "pdf_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Path"
          },
          "image_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Path"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "title": "ReportResponse",
        "description": "报告生成响应"
      },
      "SeparateRequest": {
        "properties": {
          "filepath": {
            "type": "string",
            "title": "Filepath"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "htdemucs_ft"
          },
          "two_stems": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Two Stems",
            "default": "vocals"
          }
        },
        "type": "object",
        "required": [
          "filepath"
        ],
        "title": "SeparateRequest",
        "description": "人声分离请求"
      },
      "SeparateResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "vocals_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vocals Path"
          },
          "accompaniment_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accompaniment Path"
          },
          "drums_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drums Path"
          },
          "bass_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bass Path"
          },
          "other_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Path"
          },
          "duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "model_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Used"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "title": "SeparateResponse",
        "description": "人声分离响应"
      },
...
      "UploadResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "analysis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Id"
          },
          "total_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Total Score"
          },
          "scores": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Scores"
          },
          "timbre_adjustment": {
            "type": "number",
            "title": "Timbre Adjustment",
            "default": 0.0
          },
          "level": {
            "type": "string",
            "title": "Level",
            "default": ""
          },
          "grade": {
            "type": "string",
            "title": "Grade",
            "default": ""
          },
          "advice": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Advice"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "quick"
          },
          "is_voice": {
            "type": "boolean",
            "title": "Is Voice",
            "default": true
          },
          "filepath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filepath"
          },
          "basic_info": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Basic Info"
          },
          "heuristic_dimensions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Heuristic Dimensions"
          }
        },
        "type": "object",
        "required": [
          "total_score"
        ],
        "title": "UploadResponse",
        "description": "上传分析响应 — v7.0 六维格式"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}